Reorganize conversion code
